Remove impl_col_row
